requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
rich>=13.7.0
click>=8.1.0
pyyaml>=6.0.0

# Dev
pytest>=8.0.0
